Interfaz general
Interfaz base del cliente
