--- Odyssey Cliente/Odyssey Cliente.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2036
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Odyssey Cliente", "Odyssey Cliente.csproj", "{AFBF31F5-B76C-4CB1-BB05-1E90694F8325}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AFBF31F5-B76C-4CB1-BB05-1E90694F8325}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AFBF31F5-B76C-4CB1-BB05-1E90694F8325}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AFBF31F5-B76C-4CB1-BB05-1E90694F8325}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AFBF31F5-B76C-4CB1-BB05-1E90694F8325}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2B377C3E-2BA6-4E7F-99C6-90F8192192CE}
-	EndGlobalSection
-EndGlobal
