+++ Prueba/Prueba.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2043
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Prueba", "Prueba\Prueba.csproj", "{6CB67B1B-031E-4D22-BEFE-A68C191B83F0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6CB67B1B-031E-4D22-BEFE-A68C191B83F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CB67B1B-031E-4D22-BEFE-A68C191B83F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CB67B1B-031E-4D22-BEFE-A68C191B83F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CB67B1B-031E-4D22-BEFE-A68C191B83F0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FCDBFCB7-3623-4B10-902B-A67FAED32E07}
+	EndGlobalSection
+EndGlobal
