+++ Prueba/Prueba/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // La información general de un ensamblado se controla mediante el siguiente 
 // conjunto de atributos. Cambie estos valores de atributo para modificar la información
 // asociada con un ensamblado.
-[assembly: AssemblyTitle("Odyssey Cliente")]
+[assembly: AssemblyTitle("Prueba")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Odyssey Cliente")]
+[assembly: AssemblyProduct("Prueba")]
 [assembly: AssemblyCopyright("Copyright ©  2018")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // El siguiente GUID sirve como id. de typelib si este proyecto se expone a COM.
-[assembly: Guid("afbf31f5-b76c-4cb1-bb05-1e90694f8325")]
+[assembly: Guid("6cb67b1b-031e-4d22-befe-a68c191b83f0")]
 
 // La información de versión de un ensamblado consta de los cuatro valores siguientes:
 //
